% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scrape website.R
\name{scrape_preannouncements}
\alias{scrape_preannouncements}
\title{GOV.UK statistics pre-announcements calendar scraping tool}
\usage{
scrape_preannouncements(org = NA)
}
\arguments{
\item{org}{The organisation that you wish to get pre-announcements for. This must be in the same format that the organisation appears in the URL of the pre-announcements page (https://www.gov.uk/search/research-and-statistics), when the organisation is selected as a filter. For example, "uk-health-security-agency".}
}
\value{
A .csv file, saved to your working directory.
}
\description{
Function to scrape information from the pre-announcements calendar, exporting this information into a useful spreadsheet
}
\author{
Ben Cuff
}
