% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_table.R
\name{tidy_up}
\alias{tidy_up}
\title{GOV.UK statistics pre-announcements calendar scraping tool}
\usage{
tidy_up(df)
}
\arguments{
\item{df}{The table to tidy up}
}
\description{
Internal helper function for tidying up outputs. Does not need to be run by hand.
}
\author{
Ben Cuff
}
